Let Clock return an integer

+++ src/Matthias/Polling/Clock.php
@@ -5,7 +5,7 @@ namespace Matthias\Polling;
 
 interface Clock
 {
-    public function getMicrotime(): float;
+    public function getMicrotime(): int;
 
     public function sleep(int $microseconds): void;
 }
